chore(server): enable unconvert and whitespace linters and fix issues
- Enable unconvert and whitespace linters in .golangci.yml
- Fix unconvert issues (unnecessary type conversions) in request.go, snodejs.go, and others
- Fix whitespace issues (unnecessary newlines) in multiple files
- Fix staticcheck nilness warnings in rreference.go by removing tautological checks
- Remove unused code in rfile.go and rflowv2.go

+++ packages/server/internal/api/rflowv2/rflowv2_exec.go
@@ -225,7 +225,7 @@ func (s *FlowServiceV2RPC) executeFlow(
 				ID:         execID,
 				NodeID:     status.NodeID,
 				Name:       status.Name,
-				State:      int8(status.State),
+				State:      status.State,
 				ResponseID: status.AuxiliaryID,
 			}
